[REFACTOR] 프로젝트 슬라이더 컴포넌트 분리
